requirements.txt
# Core
Flask>=3.0,<4.0
Flask-Assets>=2.1
Flask-SQLAlchemy>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
Werkzeug>=3.0,<4.0
Jinja2>=3.1,<4.0
MarkupSafe>=2.1,<3.0
itsdangerous>=2.1,<3.0
click>=8.1,<9.0

# Database
psycopg2-binary>=2.9,<3.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0
kombu>=5.3,<6.0

# WSGI Server
gunicorn>=21.0,<23.0

# Scientific
numpy>=1.26,<3.0
scipy>=1.11,<2.0

# Utilities
humanize>=4.0,<5.0
Markdown>=3.5,<4.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pyScss>=1.4,<2.0
xlrd>=2.0,<3.0
webassets>=2.0,<3.0
